refactor(tests): move tests around
